HTTP client defined in startup. One more cities added.

+++ HttpClientApp/Models/OpenWeatherMapOrg/Wind.cs
@@ -2,6 +2,6 @@
 
 public class Wind
 {
-    public int speed { get; set; }
+    public float speed { get; set; }
     public int deg { get; set; }
 }
